network-monitor/requirements.txt
anthropic
python-dotenv
discord-webhook
requests
schedule
